Return html even on error

+++ updates/update_orders.php
@@ -55,7 +55,7 @@ function update_orders() {
         $order[$i]['item_message_text'] = $item['rx_number'] ? ($item['item_message_text'] ?: '') : message_text(get_days_dispensed($item)[1], $item); //Get rid of NULL. //if not syncing to order lets provide a reason why we are not filling
         $order[$i]['days_dispensed'] = $item['days_dispensed_actual'] ?: $item['days_dispensed_default'];
         $order[$i]['qty_dispensed'] = (float) ($item['qty_dispensed_actual'] ?: $item['qty_dispensed_default']); //cast to float to get rid of .000 decimal
-        $order[$i]['refills_total'] = $item['refills_total_actual'] ?: $item['refills_total_default'];
+        $order[$i]['refills_total'] = (float) ($item['refills_total_actual'] ?: $item['refills_total_default']);
         $order[$i]['price_dispensed'] = (float) ($item['price_dispensed_actual'] ?: ($item['price_dispensed_default'] ?: 0));
       }
